moved some styles to use styles when needed
